requirements.txt
torch
torchvision
transformers
accelerate
sentencepiece
pillow
numpy
safetensors
sacrebleu
rouge-score
nltk
bert-score
